requirements.txt
Pillow==8.1.0
fire==0.4.0
numpy==1.19.5
six==1.15.0
torch===1.7.1
torchvision===0.8.2
tqdm==4.56.0
typing-extensions==3.7.4.3
